Consolidate graph utilities
